Ajuste código categorias
Tinha criado a classe ProductCard e agora listei todos os produtos de acordo com categoria que representa.
