+++ lib/models/ProductsList.dart
@@ -20,35 +20,6 @@ class Product {
   });
 }
 
-// class ProductCategory {
-//   final String name;
-//   final List<Product> products;
-
-//   ProductCategory({
-//     required this.name,
-//     required this.products,
-//   });
-// }
-
-// // Todas as categorias
-// class ProductData {
-//   static List<ProductCategory> allCategories = [
-//     ProductCategory(name: 'SmartTVs', products: TvProduct.tvProducts),
-//     ProductCategory(
-//         name: 'Eletronics', products: EletronicProduct.eletronicProducts),
-//     ProductCategory(
-//         name: 'SmartPhone', products: SmartphoneProducts.smartphoneProducts),
-//   ];
-
-//   static List<Product> getProductsByCategory(String categoryName) {
-//     final category = allCategories.firstWhere(
-//       (cat) => cat.name == categoryName,
-//       orElse: () => ProductCategory(name: "", products: []),
-//     );
-//     return category.products;
-//   }
-// }
-
 //Listar categoria SmartTvs
 List<Product> tvProducts = []..addAll(TvProduct.tvProducts);
